Implement party finder GUI enhancements and filtering
Added advanced filtering logic and party info rendering to PartyFinderGUI, including new input components and SVG icon usage. Refactored party requirements and stats handling, improved event registration for key actions, and updated helper and data classes for consistency. Also added new image assets and fixed minor UI and logic bugs.

+++ src/main/kotlin/net/sbo/mod/utils/Helper.kt
@@ -154,7 +154,7 @@ object Helper {
         }
     }
 
-    fun matchDianaKillsToColor(kills: Long): String {
+    fun matchDianaKillsToColor(kills: Int): String {
         return when {
             kills >= 200_000 -> "§6${formatNumber(kills, true)}"
             kills >= 150_000 -> "§e${formatNumber(kills, true)}"
